+++ xprof/Makefile.am
@@ -20,8 +20,9 @@ xprof_utils.hpp: $(top_srcdir)/utils/xprof_utils.hpp
 %.pb.h %.pb.cc: %.proto
 	$(PROTOC) -I=$(srcdir) --cpp_out=$(builddir) $<
 
-btx_tally_upstream_model.yaml btx_timeline_upstream_model.yaml: btx_interval_model.yaml
-	cp $< $(builddir)/$@
+btx_timeline_upstream_model.yaml btx_tally_upstream_model.yaml &: btx_interval_model.yaml
+	cp $< $(builddir)/btx_timeline_upstream_model.yaml
+	cp $< $(builddir)/btx_tally_upstream_model.yaml
 
 %/metababel/metababel.h %/metababel/btx_component.h %/metababel/btx_component.c %/metababel/btx_upstream.h %/metababel/btx_upstream.c %/btx_main.c: %_upstream_model.yaml %_params.yaml
 	$(METABABEL) -u $(builddir)/$*_upstream_model.yaml -t SINK -o $* -p $* -c $* --params $(srcdir)/$*_params.yaml
@@ -65,7 +66,6 @@ nodist_libXTimeline_la_SOURCES = \
 
 # File generated
 libXTally_la_SOURCES = \
-	tally_utils.hpp \
 	tally.hpp \
 	tally.cpp \
 	my_demangle.c \
